api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
transformers==4.29.2
ftfy==6.1.1
faiss-cpu==1.7.4
numpy==1.24.3
opencv-python==4.7.0.72
python-dotenv==1.0.0
gunicorn==20.1.0 